Refactor to Timmy package
