+++ src/main/java/Timmy/Deadline.java
@@ -1,19 +1,19 @@
+package Timmy;
+
 import Exceptions.TimmyDateParsingException;
 
 import java.time.LocalDate;
 import java.time.format.DateTimeParseException;
 import java.time.format.DateTimeFormatter;
 
-public class Event extends Task {
-    protected final LocalDate start;
+public class Deadline extends Task {
     protected final LocalDate end;
     protected final DateTimeFormatter INPUT_DATE_FORMAT = DateTimeFormatter.ofPattern("d/M/yyyy");
     protected final DateTimeFormatter OUTPUT_DATE_FORMAT = DateTimeFormatter.ofPattern("MMM dd yyyy");
 
-    Event(String desc, String start, String end) throws DateTimeParseException {
+    Deadline(String desc, String end) throws TimmyDateParsingException {
         super(desc);
         try {
-            this.start = LocalDate.parse(start, INPUT_DATE_FORMAT);
             this.end = LocalDate.parse(end, INPUT_DATE_FORMAT);
         } catch (DateTimeParseException e) {
             throw new TimmyDateParsingException();
@@ -21,13 +21,11 @@ public class Event extends Task {
     }
 
     public String toCompleteString() {
-        return "[E]" + super.toStringWithStatusIcon()
-                + " (from: " + start.format(OUTPUT_DATE_FORMAT)
-                + " to: " + end.format(OUTPUT_DATE_FORMAT) + ")";
+        return "[D]" + super.toStringWithStatusIcon()
+                + " (by: " + end.format(OUTPUT_DATE_FORMAT) + ")";
     }
 
     public String toFileString() {
-        return "E | " + (isDone ? "1" : "0") + " | " + super.toString()
-                + " | " + start.format(OUTPUT_DATE_FORMAT) + " | " + end.format(INPUT_DATE_FORMAT);
+        return "D | " + (isDone ? "1" : "0") + " | " + super.toString() + " | " + end.format(INPUT_DATE_FORMAT);
     }
 }
