+++ src/main/java/Timmy/Event.java
@@ -1,17 +1,21 @@
+package Timmy;
+
 import Exceptions.TimmyDateParsingException;
 
 import java.time.LocalDate;
 import java.time.format.DateTimeParseException;
 import java.time.format.DateTimeFormatter;
 
-public class Deadline extends Task {
+public class Event extends Task {
+    protected final LocalDate start;
     protected final LocalDate end;
     protected final DateTimeFormatter INPUT_DATE_FORMAT = DateTimeFormatter.ofPattern("d/M/yyyy");
     protected final DateTimeFormatter OUTPUT_DATE_FORMAT = DateTimeFormatter.ofPattern("MMM dd yyyy");
 
-    Deadline(String desc, String end) throws TimmyDateParsingException {
+    Event(String desc, String start, String end) throws DateTimeParseException {
         super(desc);
         try {
+            this.start = LocalDate.parse(start, INPUT_DATE_FORMAT);
             this.end = LocalDate.parse(end, INPUT_DATE_FORMAT);
         } catch (DateTimeParseException e) {
             throw new TimmyDateParsingException();
@@ -19,11 +23,13 @@ public class Deadline extends Task {
     }
 
     public String toCompleteString() {
-        return "[D]" + super.toStringWithStatusIcon()
-                + " (by: " + end.format(OUTPUT_DATE_FORMAT) + ")";
+        return "[E]" + super.toStringWithStatusIcon()
+                + " (from: " + start.format(OUTPUT_DATE_FORMAT)
+                + " to: " + end.format(OUTPUT_DATE_FORMAT) + ")";
     }
 
     public String toFileString() {
-        return "D | " + (isDone ? "1" : "0") + " | " + super.toString() + " | " + end.format(INPUT_DATE_FORMAT);
+        return "E | " + (isDone ? "1" : "0") + " | " + super.toString()
+                + " | " + start.format(OUTPUT_DATE_FORMAT) + " | " + end.format(INPUT_DATE_FORMAT);
     }
 }
