+++ src/main/java/Timmy/Task.java
@@ -1,3 +1,5 @@
+package Timmy;
+
 abstract public class Task {
     protected final String description;
     protected boolean isDone;
