+++ src/main/java/Timmy/TaskList.java
@@ -1,6 +1,7 @@
+package Timmy;
+
 import Exceptions.TimmyTaskListOutOfBoundsException;
 
-import java.lang.reflect.Array;
 import java.util.ArrayList;
 
 public class TaskList {
