+++ src/main/java/Timmy/Ui.java
@@ -1,3 +1,5 @@
+package Timmy;
+
 import java.util.ArrayList;
 
 public class Ui {
@@ -14,7 +16,7 @@ public class Ui {
     }
 
     public void showWelcomeMessage() {
-        showMessage(CHATBOT_TEXT_OFFSET + "Hello! I'm Timmy\n"
+        showMessage(CHATBOT_TEXT_OFFSET + "Hello! I'm Timmy.Timmy\n"
                 + CHATBOT_TEXT_OFFSET + "What can I do for you?");
     }
