+++ src/main/java/Timmy/ValidCommand.java
@@ -1,3 +1,5 @@
+package Timmy;
+
 public enum ValidCommand {
     BYE,
     LIST,
