+++ text-ui-test/EXPECTED.TXT
@@ -1,5 +1,5 @@
     ____________________________________________________________
-     Hello! I'm Timmy
+     Hello! I'm Timmy.Timmy
      What can I do for you?
     ____________________________________________________________
     ____________________________________________________________
